Resolved conflicts in:
- parallel.rs: Combined declares_sorted_output() with RowEstimate::Known check
- joinscan/mod.rs: Adopted collect_join_sources approach with Vec return type
- joinscan/planning.rs: Kept main's structure with sort_order extraction

+++ pg_search/tests/pg_regress/sql/aggregate-groupby-conflict.sql
@@ -10,7 +10,7 @@ CREATE TABLE groupby_conflict_test (
     title TEXT,
     category TEXT,
     rating INTEGER,
-    price NUMERIC,
+    price FLOAT,
     views INTEGER
 );
